API SISGIU/requirements.txt
Django==2.0
django-filter==1.1.0
djangorestframework==3.7.7
Markdown==2.6.10
Pillow==5.0.0
psycopg2==2.7.3.2